[#51] 자동 위치를 통한 코디 수동 생성 api 적용
#51

+++ Codive/Features/Home/Domain/Entities/HomeCodiEntity.swift
@@ -31,7 +31,7 @@ struct TodayDailyCodi {
 }
 
 struct CodiPayload {
-    let clothId: Int
+    let clothId: Int64
     let locationX: Double
     let locationY: Double
     let ratio: Double
